# packages/api-client/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "REHAU Nea Smart 2 — Bridge API",
    "description": "REST + SSE access to a REHAU Nea Smart 2.0 base station.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/installer/calibration": {
      "get": {
        "tags": [
          "installer"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "outdoor": {
                      "type": "number"
                    },
                    "rooms": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "zone": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 63
                          },
                          "tempOffset": {
                            "type": "number",
                            "minimum": -5,
                            "maximum": 5
                          },
                          "humidityOffset": {
                            "type": "integer",
                            "minimum": -25,
                            "maximum": 25
                          }
                        },
                        "required": [
                          "zone",
                          "tempOffset",
                          "humidityOffset"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "lastUpdatedAt": {
                          "type": "string",
                          "description": "ISO timestamp"
                        }
                      },
                      "required": [
                        "lastUpdatedAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "outdoor",
                    "rooms",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "installer"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "outdoor": {
                    "type": "number",
                    "minimum": -10,
                    "maximum": 10
                  },
                  "rooms": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "zone": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 63
                        },
                        "tempOffset": {
                          "type": "number",
                          "minimum": -5,
                          "maximum": 5
                        },
                        "humidityOffset": {
                          "type": "integer",
                          "minimum": -25,
                          "maximum": 25
                        }
                      },
                      "required": [
                        "zone",
                        "tempOffset",
                        "humidityOffset"
                      ],
                      "additionalProperties": false
                    }
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "outdoor": {
                      "type": "number"
                    },
                    "rooms": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "zone": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 63
                          },
                          "tempOffset": {
                            "type": "number",
                            "minimum": -5,
                            "maximum": 5
                          },
                          "humidityOffset": {
                            "type": "integer",
                            "minimum": -25,
                            "maximum": 25
                          }
                        },
                        "required": [
                          "zone",
                          "tempOffset",
                          "humidityOffset"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "lastUpdatedAt": {
                          "type": "string",
                          "description": "ISO timestamp"
                        }
                      },
                      "required": [
                        "lastUpdatedAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "outdoor",
                    "rooms",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/installer/io": {
      "get": {
        "tags": [
          "installer"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "master": {
                      "type": "object",
                      "properties": {
                        "rz": {
                          "type": "array",
                          "items": {
                            "type": "number"
                          }
                        },
                        "relay": {
                          "type": "array",
                          "items": {
                            "type": "number"
                          }
                        },
                        "di": {
                          "type": "array",
                          "items": {
                            "type": "number"
                          }
                        }
                      },
                      "required": [
                        "rz",
                        "relay",
                        "di"
                      ],
                      "additionalProperties": false
                    },
                    "umodules": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "properties": {
                          "relay": {
                            "type": "array",
                            "items": {
                              "type": "number"
                            }
                          },
                          "di": {
                            "type": "array",
                            "items": {
                              "type": "number"
                            }
                          },
                          "aiC": {
                            "type": "array",
                            "items": {
                              "type": "number",
                              "nullable": true
                            }
                          },
                          "aoPct": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "relay",
                          "di",
                          "aiC",
                          "aoPct"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "master",
                    "umodules"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/installer/diagnostics/uptime": {
      "get": {
        "tags": [
          "installer"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "years": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "days": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "hours": {
                      "type": "integer",
                      "minimum": 0
                    }
                  },
                  "required": [
                    "years",
                    "days",
                    "hours"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/installer/diagnostics/topology": {
      "get": {
        "tags": [
          "installer"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "baseModules": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "rModules": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "uModules": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "rooms": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "mixedCircuits": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "dehumidifiers": {
                      "type": "integer",
                      "minimum": 0
                    }
                  },
                  "required": [
                    "baseModules",
                    "rModules",
                    "uModules",
                    "rooms",
                    "mixedCircuits",
                    "dehumidifiers"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/installer/curve": {
      "get": {
        "tags": [
          "installer"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "slopeNormal": {
                      "type": "number"
                    },
                    "slopeAbsent": {
                      "type": "number"
                    },
                    "startNormal": {
                      "type": "number"
                    },
                    "startAbsent": {
                      "type": "number"
                    },
                    "reductionK": {
                      "type": "number"
                    },
                    "minFlowNormalC": {
                      "type": "number"
                    },
                    "minFlowAbsentC": {
                      "type": "number"
                    },
                    "maxFlowNormalC": {
                      "type": "number"
                    },
                    "maxFlowAbsentC": {
                      "type": "number"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "lastUpdatedAt": {
                          "type": "string",
                          "description": "ISO timestamp"
                        }
                      },
                      "required": [
                        "lastUpdatedAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "slopeNormal",
                    "slopeAbsent",
                    "startNormal",
                    "startAbsent",
                    "reductionK",
                    "minFlowNormalC",
                    "minFlowAbsentC",
                    "maxFlowNormalC",
                    "maxFlowAbsentC",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ]
}